fix tests and make test float for all functional tests
