+++ tests/Functional/Schema/SchemaManagerFunctionalTestCase.php
@@ -27,7 +27,9 @@ use Doctrine\DBAL\Types\BlobType;
 use Doctrine\DBAL\Types\DateTimeType;
 use Doctrine\DBAL\Types\DateType;
 use Doctrine\DBAL\Types\DecimalType;
+use Doctrine\DBAL\Types\FloatType;
 use Doctrine\DBAL\Types\IntegerType;
+use Doctrine\DBAL\Types\RealFloatType;
 use Doctrine\DBAL\Types\StringType;
 use Doctrine\DBAL\Types\TextType;
 use Doctrine\DBAL\Types\TimeType;
@@ -857,6 +859,20 @@ abstract class SchemaManagerFunctionalTestCase extends FunctionalTestCase
 
         self::assertTrue($created->hasColumn('binarydata'));
         self::assertInstanceOf(BlobType::class, $created->getColumn('binarydata')->getType());
+    }
+
+    public function testListTableFloatTypeColumns(): void
+    {
+        $table = new Table('test_float_table');
+        $table->addColumn('col_float', Types::FLOAT);
+        $table->addColumn('col_real_float', Types::REAL);
+
+        $this->dropAndCreateTable($table);
+
+        $columns = $this->schemaManager->listTableColumns('test_float_table');
+
+        self::assertInstanceOf(FloatType::class, $columns['col_float']->getType());
+        self::assertInstanceOf(RealFloatType::class, $columns['col_real_float']->getType());
     }
 
     /** @param mixed[] $data */
